# src/swagger/output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Canban API",
    "description": "Documentation automatically generated by the <b>swagger-autogen</b> module.",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/users/{id}": {
      "get": {
        "tags": [
          "Users"
        ],
        "description": "getting a user by id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "user data",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      }
    },
    "/users/create": {
      "post": {
        "tags": [
          "Users"
        ],
        "description": "create a new user",
        "parameters": [
          {
            "name": "data",
            "in": "body",
            "description": "new user",
            "required": true,
            "schema": {
              "$ref": "#/definitions/NewUser"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "created user",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      }
    },
    "/users/update/{id}": {
      "put": {
        "tags": [
          "Users"
        ],
        "description": "update an existing user",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "data",
            "in": "body",
            "description": "user fields to update",
            "required": true,
            "schema": {
              "$ref": "#/definitions/NewUser"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "updated user",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      }
    },
    "/users/delete/{id}": {
      "delete": {
        "tags": [
          "Users"
        ],
        "description": "user removal",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "removal success",
            "schema": {
              "$ref": "#/definitions/Success"
            }
          }
        }
      }
    },
...
    "/statuses/{id}": {
      "get": {
        "tags": [
          "Statuses"
        ],
        "description": "getting a status by id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "status",
            "schema": {
              "$ref": "#/definitions/Status"
            }
          }
        }
      }
    },
    "/statuses/create": {
      "post": {
        "tags": [
          "Statuses"
        ],
        "description": "create a new status",
        "parameters": [
          {
            "name": "data",
            "in": "body",
            "description": "new status",
            "required": true,
            "schema": {
              "$ref": "#/definitions/NewStatus"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "created status",
            "schema": {
              "$ref": "#/definitions/Status"
            }
          }
        }
      }
    },
    "/statuses/update/{id}": {
      "put": {
        "tags": [
          "Statuses"
        ],
        "description": "update an existing status",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "data",
            "in": "body",
            "description": "status fields to update",
            "required": true,
            "schema": {
              "$ref": "#/definitions/NewStatus"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "updated status",
            "schema": {
              "$ref": "#/definitions/Statuses"
            }
          }
        }
      }
    },
    "/statuses/delete/{id}": {
      "delete": {
        "tags": [
          "Statuses"
        ],
        "description": "status removal",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "removal success",
            "schema": {
              "$ref": "#/definitions/Success"
            }
          }
        }
      }
    },
...
    "/tasks/{id}": {
      "get": {
        "tags": [
          "Tasks"
        ],
        "description": "getting a task by id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "task",
            "schema": {
              "$ref": "#/definitions/Task"
            }
          }
        }
      }
    },
    "/tasks/create": {
      "post": {
        "tags": [
          "Tasks"
        ],
        "description": "create a new task",
        "parameters": [
          {
            "name": "data",
            "in": "body",
            "description": "new task",
            "required": true,
            "schema": {
              "$ref": "#/definitions/NewTask"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "created task",
            "schema": {
              "$ref": "#/definitions/Task"
            }
          }
        }
      }
    },
    "/tasks/update/{id}": {
      "put": {
        "tags": [
          "Tasks"
        ],
        "description": "update an existing task",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "data",
            "in": "body",
            "description": "task fields to update",
            "required": true,
            "schema": {
              "$ref": "#/definitions/NewTask"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "updated task",
            "schema": {
              "$ref": "#/definitions/Task"
            }
          }
        }
      }
    },
...
    "/tasks/delete/{id}": {
      "delete": {
        "tags": [
          "Tasks"
        ],
        "description": "task removal",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "removal success",
            "schema": {
              "$ref": "#/definitions/Success"
            }
          }
        }
      }
    }
  },
  "definitions": {
...
    "Success": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean",
          "example": true
        }
      }
    },
...
    "Task": {
      "type": "object",
      "properties": {
        "_id": {
          "type": "string",
          "example": "61e7de469343e960064dce9e"
        },
        "name": {
          "type": "string",
          "example": "Super Task Name 1"
        },
        "description": {
          "type": "string",
          "example": "Super Task Description"
        },
        "creator": {
          "type": "string",
          "example": "61e5364f347b2493b8a42782"
        },
        "assignedTo": {
          "type": "string",
          "example": "61e5364f347b2493b8a42782"
        },
        "boardId": {
          "type": "string",
          "example": "61e65d624327f14b2f7bf00c"
        },
        "statusId": {
          "type": "string",
          "example": "61e65dfa4327f14b2f7bf00e"
        },
        "createdAt": {
          "type": "string",
          "example": "1970-01-20T00:16:25.670Z"
        },
        "updatedAt": {
          "type": "string",
          "example": "1970-01-20T00:16:25.670Z"
        },
        "elapsedTime": {
          "type": "number",
          "example": 1641005765247
        },
        "id": {
          "type": "string",
          "example": "61e7de469343e960064dce9e"
        }
      }
    },
...
    "NewTask": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "Super Task Name 1"
        },
        "description": {
          "type": "string",
          "example": "Super Task Description"
        },
        "assignedTo": {
          "type": "string",
          "example": "61e5364f347b2493b8a42782"
        },
        "boardId": {
          "type": "string",
          "example": "61e65d624327f14b2f7bf00c"
        },
        "statusId": {
          "type": "string",
          "example": "61e65dfa4327f14b2f7bf00e"
        }
      }
    },
    "Status": {
      "type": "object",
      "properties": {
        "_id": {
          "type": "string",
          "example": "61e65e4f5914c628f9ae5656"
        },
        "name": {
          "type": "string",
          "example": "In progress"
        },
        "boardIds": {
          "type": "array",
          "example": [
            "61e65d624327f14b2f7bf00c"
          ],
          "items": {
            "type": "string"
          }
        }
      }
    },
    "Statuses": {
      "type": "array",
      "items": {
        "$ref": "#/definitions/Status"
      }
    },
    "NewStatus": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "New status name"
        },
        "boardId": {
          "type": "string",
          "example": "61e65d624327f14b2f7bf00c"
        }
      }
    },
    "User": {
      "type": "object",
      "properties": {
        "_id": {
          "type": "string",
          "example": "61e5374b347b2493b8a42783"
        },
        "name": {
          "type": "string",
          "example": "John Doe"
        },
        "email": {
          "type": "string",
          "example": "akrays@aspirity.com"
        }
      }
    },
...
    "NewUser": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "John Doe"
        },
        "password": {
          "type": "string",
          "example": "12345678"
        },
        "email": {
          "type": "string",
          "example": "john@example.com"
        },
        "projectsId": {
          "type": "array",
          "example": [
            "61e65d624327f14b2f7bf00c"
          ],
          "items": {
            "type": "string"
          }
        }
      }
    }
  }
}